Don't scroll before layout has happened.
Instead scrolling is deferred until layout has been performed. This avoids the
use of a Handler.post() in onPostCreate(), which means that additional show*()
calls performed by a sub-class of SlidingActivity in onPostCreate() work as
expected.

Also derive SlidingMenu from FrameLayout instead of RelativeLayout as there
seems to be no reason to use RelativeLayout.