Avoid touch lockup due to mQuickReturn == true
This can happen when a DOWN event is triggered in the quickreturn
area, and the finger then moved out of that area without triggering
a drag. If the menu is then closed in some other way (e.g. back key)
the CustomViewAbove remains in mQuickReturn == true state and blocks
all touches to the main content.